temporary report section fix
The go_xpost_pull was timing out while the filter processed report
subsections for xposting. This is bad.

We also haven't quite figured out how to display the report subsections
cleanly, causing some frustration there.

The temporary answer is to concatenate all the subsection content into
the top-level report content before xposting. That preserves overall
findability for the report (though not the individual sections) and
solves the timeout and display problems we've been struggling with.

Leaving the relevent pieces in place and commented out so we can return
to them and develop a better solution.