Compare encoded signatures so we are comparing 1 to 1
Some combinations of action/post_id/filter/source/secret cause the signature to include characters that get URL encoded during the construction of the hash. Example: `==` becomes `%3D%3D`.  When analyzing the signatures on the receiving end (whether as a ping or a pull request), the unencoded signature was getting compared with an encoded one causing the signature check to fail.

95% of the time there weren't any special characters. I did some spot checking on posts that were never xposted and the resulting signatures all had `==` as part of them.